backend/requirements.txt
python-dotenv
aiohttp
asyncio
websockets